# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Sales Server",
    "description": "Documentation API. https://github.com/danielschmitz/sales-server <br/><br/>",
    "version": "1.0",
    "contact": {
      "name": "Daniel Schmitz",
      "email": "danieljfa@gmail.com",
      "url": "https://github.com/danielschmitz"
    }
  },
  "basePath": "/api",
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "scheme": "bearer",
      "in": "header"
    }
  },
  "paths": {
...
    "/user/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get a User",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User",
            "schema": {
              "$ref": "#/definitions/UserResult"
            }
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "🔒️ Edit a User",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "user",
            "in": "body",
            "description": "User Data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated successfully.",
            "schema": {
              "$ref": "#/definitions/UserResult"
            }
          },
          "422": {
            "description": "Invalid input"
          },
          "500": {
            "description": "There is already a user with that email"
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "🔒️ Delete a User",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted"
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/user": {
      "post": {
        "tags": ["Users"],
        "summary": "Create a new User",
        "description": "",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User Data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User registered successfully.",
            "schema": {
              "$ref": "#/definitions/UserResult"
            }
          },
          "422": {
            "description": "Invalid input"
          },
          "500": {
            "description": "There is already a email with that name"
          }
        }
      }
    },
    "/user/changePassword/{id}": {
      "put": {
        "tags": ["Users"],
        "summary": "🔒️ Change a password",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "user",
            "in": "body",
            "description": "User Data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string",
                  "example": "123@456"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password updated successfully.",
            "schema": {
              "$ref": "#/definitions/UserResult"
            }
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/supplier/{id}": {
      "get": {
        "tags": ["Suppliers"],
        "summary": "Get a Supplier",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier",
            "schema": {
              "$ref": "#/definitions/SupplierResult"
            }
          },
          "404": {
            "description": "Supplier not found"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": ["Suppliers"],
        "summary": "🔒️ Edit a Supplier",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "supplier",
            "in": "body",
            "description": "Supplier Data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Supplier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier registered successfully.",
            "schema": {
              "$ref": "#/definitions/SupplierResult"
            }
          },
          "422": {
            "description": "Invalid input"
          },
          "500": {
            "description": "There is already a supplier with that name"
          }
        }
      },
      "delete": {
        "tags": ["Suppliers"],
        "summary": "🔒️ Delete a Supplier",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier deleted"
          },
          "404": {
            "description": "Supplier not found"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/supplier": {
      "post": {
        "tags": ["Suppliers"],
        "summary": "🔒️ Create a new Supplier",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "supplier",
            "in": "body",
            "description": "Supplier Data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Supplier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier registered successfully.",
            "schema": {
              "$ref": "#/definitions/SupplierResult"
            }
          },
          "422": {
            "description": "Invalid input"
          },
          "500": {
            "description": "There is already a supplier with that name"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UserResult": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "User Name"
        },
        "email": {
          "type": "string",
          "example": "user@mail.com"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "User Name"
        },
        "email": {
          "type": "string",
          "example": "user@mail.com"
        },
        "password": {
          "type": "string",
          "example": "123@456"
        }
      }
    },
    "SupplierResult": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "companyName": {
          "type": "string",
          "example": "Supplier Company Name"
        },
        "contactName": {
          "type": "string",
          "example": "Supplier Contact Name"
        },
        "contactEmail": {
          "type": "string",
          "example": "supplier@email.com"
        },
        "address_id": {
          "type": "number",
          "example": 10
        },
        "address": {
          "type": "object",
          "properties": {
            "id": {
              "type": "number",
              "example": 10
            },
            "street": {
              "type": "string",
              "example": "Street Address"
            },
            "city": {
              "type": "string",
              "example": "City"
            },
            "region": {
              "type": "string",
              "example": "Region"
            },
            "postalCode": {
              "type": "string",
              "example": "Postal Code"
            },
            "country": {
              "type": "string",
              "example": "Country"
            },
            "phone": {
              "type": "string",
              "example": "Phone"
            }
          }
        }
      }
    },
...
    "Supplier": {
      "type": "object",
      "properties": {
        "companyName": {
          "type": "string",
          "example": "Supplier Company Name"
        },
        "contactName": {
          "type": "string",
          "example": "Supplier Contact Name"
        },
        "contactEmail": {
          "type": "string",
          "example": "supplier@email.com"
        },
        "address": {
          "type": "object",
          "properties": {
            "street": {
              "type": "string",
              "example": "Street Address"
            },
            "city": {
              "type": "string",
              "example": "City"
            },
            "region": {
              "type": "string",
              "example": "Region"
            },
            "postalCode": {
              "type": "string",
              "example": "Postal Code"
            },
            "country": {
              "type": "string",
              "example": "Country"
            },
            "phone": {
              "type": "string",
              "example": "Phone"
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}